Implement OAuth services for integration with Google Sheets API.
Includes the fetching and exchange of an authorization code for an
access token. This access token is then used to query the Sheets API.
Some logic to use a refresh token is also included. This allows the
server to automatically exchange an _expired_ access token for a new one
without having the user go through the (external) Google authorization
process again.