Create Integration, DBIntegration + IntegrationBlock types
Start to implement some new GraphQL and DB-related types to facilitate
adding of new Integrations to a portal.
Also, start to add logic allowing for the creation of an
IntegrationBlock. This will be associated with one Integration and its
relevant dimensions.
The workflow also adheres to the "block configuration" workflow as
discussed. This mainly involves initializing the row and column
`Dimension`s, creating the `IntegrationBlock` as well as the
new `GoogleSheetsCell`, which will be responsible for making the call to
the Sheets API (to fetch the value, given the row and column
dimensions).